chore(suite-native): Mobile Trade: types renaming, generic trade input
