+++ suite-native/forms/src/Form.tsx
@@ -1,7 +1,7 @@
 import { ReactNode, createContext } from 'react';
 import { FieldValues, UseFormReturn } from 'react-hook-form';
 
-export type { FieldValues, UseFormReturn } from 'react-hook-form';
+export type { FieldPath, FieldValues, UseFormReturn } from 'react-hook-form';
 
 export interface FormProps<TFieldValues extends FieldValues> {
     children?: ReactNode;
