+++ suite-native/module-trading/src/components/buy/__tests__/BuyAlert.comp.test.tsx
@@ -6,11 +6,11 @@ import {
 } from '@suite-native/test-utils';
 
 import { useBuyForm } from '../../../hooks/buy/useBuyForm';
-import { TradingBuyForm } from '../../../types';
+import { BuyForm } from '../../../types/buy';
 import { BuyAlert } from '../BuyAlert';
 
 describe('BuyAlert', () => {
-    let form: TradingBuyForm;
+    let form: BuyForm;
 
     const renderFormHook = () => renderHookWithStoreProviderAsync(() => useBuyForm());
