+++ suite-native/module-trading/src/hooks/general/__tests__/useFavouriteAssetsSectionList.test.tsx
@@ -8,7 +8,7 @@ import {
 } from '@suite-native/test-utils';
 
 import { adaAsset, btcAsset, usdcAsset } from '../../../__fixtures__/tradeableAssets';
-import { TradeableAsset } from '../../../types';
+import { TradeableAsset } from '../../../types/general';
 import { useFavouriteAssetsSectionList } from '../useFavouriteAssetsSectionList';
 
 describe('useFavouriteAssetsSectionList', () => {
