+++ suite-native/module-trading/src/hooks/general/useSheetControls.ts
@@ -1,15 +1,24 @@
 import { useCallback } from 'react';
 
-import { useBottomSheetControls } from './useBottomSheetControls';
-import { TradingBuyForm, TradingBuyFormValues } from '../../types';
+import type { FieldPath, FieldValues, UseFormReturn } from '@suite-native/forms';
 
-export const useSheetControls = <Key extends keyof TradingBuyFormValues>(
-    { setValue, watch }: TradingBuyForm,
+import { useBottomSheetControls } from './useBottomSheetControls';
+
+export type SheetControls<TFieldValues extends FieldValues, Key extends FieldPath<TFieldValues>> = {
+    selectedValue: TFieldValues[Key];
+    setSelectedValue: (value: TFieldValues[Key]) => void;
+} & ReturnType<typeof useBottomSheetControls>;
+
+export const useSheetControls = <
+    TFieldValues extends FieldValues,
+    Key extends FieldPath<TFieldValues>,
+>(
+    { setValue, watch }: UseFormReturn<TFieldValues>,
     key: Key,
-) => {
+): SheetControls<TFieldValues, Key> => {
     const bottomSheetControls = useBottomSheetControls();
 
-    const selectedValue = watch(key);
+    const selectedValue = watch(key as FieldPath<TFieldValues>);
 
     const setSelectedValue = useCallback(
         (value: typeof selectedValue) => setValue(key, value),
