+++ suite-native/module-trading/src/index.ts
@@ -1,5 +1,4 @@
 export * from './navigation/TradingStackNavigator';
 export * from './tradingSlice';
 export * from './selectors/commonSelectors';
-export * from './types';
 export * from './screens/TradingWebViewScreen';
