+++ suite-native/module-trading/src/utils/general/tradeableAssetUtils.ts
@@ -4,7 +4,8 @@ import { cryptoIdToSymbol, isCryptoIdForNativeToken, parseCryptoId } from '@suit
 import { NetworkSymbolExtended } from '@suite-common/wallet-config';
 import { TokenAddress } from '@suite-common/wallet-types';
 
-import { TradeableAsset, TradingBuyForm } from '../../types';
+import { BuyForm } from '../../types/buy';
+import { TradeableAsset } from '../../types/general';
 
 const FAVOURITE_NETWORKS_PRIORITY_ORDER = ['bitcoin', 'ethereum', 'litecoin', 'cardano', 'solana'];
 
@@ -43,7 +44,7 @@ export const coinInfoToTradeableAsset = (
     };
 };
 
-export const getSelectedSymbolFromBuyForm = (form: TradingBuyForm) => {
+export const getSelectedSymbolFromBuyForm = (form: BuyForm) => {
     const cryptoId = form.watch('asset')?.cryptoId;
 
     return cryptoId ? cryptoIdToSymbol(cryptoId) : undefined;
